// src/app/models/insurer.ts
export class Insurer {
  _id: string;
  brandid: number;
  brand: string;
  productid: number;
  product: string;
  producttypeid: number;
  producttype: string;
  description: string;
  street: string;
  house_number: number;
  house_number_addition: string;
  zipcode: string;
  city: string;
  postaddress: string;
  telephone_number: string;
  emergency_number: string;
  whatsapp: string;
  aplaza_account: string;
  logo: string;
  facebook: string;
  twitter: string;
  url: string;
  email: string;
  response_rate: number;
  value_for_money_rating: number;
  ease_of_arranging: number;
  total_rating: number;
  ease_of_comunication: number;
  travel_service_insurance: number;
  car_service_insurance: number;
  price_insurance: number;
  reviews_insurance: number;
  travel_service_user_preference: number;
  car_service_user_preference: number;
  price_user_preference: number;
  reviews_user_preference: number;
  supported: boolean;
  phone_communication: boolean;
  policy_type: string;
  reviews: number;
  rating_scale: number;
  _identifier: string;
  _version: string;
  _current: boolean;
  _original: boolean;
  _created: {
    sec: number,
    usec: number
  };
}